Cloud Functions/gcs_to_ga/requirements.txt
google-api-python-client==1.7.4
google-cloud-storage==1.13.0
oauth2client==4.1.3